+++ NEWS.md
@@ -14,6 +14,11 @@ MINOR CHANGES
   - All `grouped_` functions had argument `title.prefix` that defaulted to
     `"Group"`. It now instead defaults to `NULL`, in which case the prefix will
     variable name for `grouping.var` argument.
+  - To accommodate non-paramtric tests, `subtitle_template` function can now
+    work with `parameter = NULL`.
+  - For `ggbetweenstats`, details contained in the subtitle for non-parametric
+    test are modified. It now uses Hodges–Lehmann–Sen estimator as effect size.
+    This removes `coin` from dependencies.
   - `ggbetweenstats` and its `grouped_` variant gain a new argument
     `axes.range.restrict` (which defaults to `FALSE`). This restricts `y`-axes
     limits to minimum and maximum of `y` variable. This is what these functions